updating scrapers, fetchproductdata

+++ src/models/price.js
@@ -24,7 +24,7 @@ export default (sequelize) => {
                 allowNull: false,
             },
             price: {
-                type: DataTypes.FLOAT,
+                type: DataTypes.DECIMAL(10, 2), // 10 digits in total, 2 after decimal
                 allowNull: false,
             },
             currency: {
